[Iteration 17] - [Amit] Bug fixes as per tracker
This check in has the following changes for the bugs mentioned in the tracker:
7	Edit Purchase Order	The item changed in case of only 1 item should not be added as another item record, it should replace the existing items
Solution: create unique order item sequence for joining

9	Edit Supplier	Retrieved a Supplier, deleted an Address
Then added phone 2 field value
Then click on Save.

At this point, the save has failed with error created_by cannot be null.
This is happening when the app is trying to create the supplier account.

The fix should be that the supplier account should not be created again in edit flow
10	Edit Purchase Order	When the changes are saved, the action buttons were not displayed correctly
Add the Purchase order management button as well
11	Manage Price Event	The bulk Delete and approve fails, test the functionality and fix it
12	Create Inventory Adjustment	The change of saved item should not create a new record rather it should update the old item record

Also, the old DB changes were synched with the DB migration scripts